feat: add geocoding to store data from public API
- 공공기관 (한국소비자원) API를 통해 매장, 상품 데이터를 불러옴
- vWorld 지오코딩을 통해 매장에 위도/경도 좌표를 추가함

+++ migrations/20251027120247_create_stores_table.sql
@@ -6,8 +6,8 @@ CREATE TABLE IF NOT EXISTS stores(
     post_no VARCHAR(10),
     plmk_addr VARCHAR(255) NOT NULL,
     road_addr VARCHAR(255) NOT NULL,
-    x_coord NUMERIC(18,15),
-    y_coord NUMERIC(18,15),
+    x_coord DOUBLE PRECISION,
+    y_coord DOUBLE PRECISION,
     created_at TIMESTAMP DEFAULT NOW(),
     updated_at TIMESTAMP DEFAULT NOW()
 );
